requirements.txt
dj-database-url==0.5.0
Django==3.2
django-auditlog==2.0.0
django-cors-headers==3.10.1
django-easy-audit==1.3.2
djangorestframework==3.13.1
django-environ==0.9.0
django-rest-swagger==2.2.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
pyjwt==1.7.1
requests==2.21.0
wincertstore==0.2


